requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
minio==7.2.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
click==8.1.7
requests==2.31.0
